Add files via upload
Updated the file as the final "disclaimer prompt" was not shown at the top of the script, but now it is :)

This is a 3rd voice that speaks a specific sentence after the conversation between the 2 LLMs has ended. This can be used for a final message. Edit it when edit the other prompts at the top of the script.